301 create relation between user and database (#302)

+++ backend/src/main/java/com/janbabak/noqlbackend/service/database/DatabaseEntityService.java
@@ -1,16 +1,20 @@
 package com.janbabak.noqlbackend.service.database;
 
 import com.janbabak.noqlbackend.dao.repository.DatabaseRepository;
+import com.janbabak.noqlbackend.dao.repository.UserRepository;
 import com.janbabak.noqlbackend.error.exception.DatabaseConnectionException;
 import com.janbabak.noqlbackend.error.exception.DatabaseExecutionException;
 import com.janbabak.noqlbackend.error.exception.EntityNotFoundException;
 import com.janbabak.noqlbackend.model.database.*;
 import com.janbabak.noqlbackend.model.entity.Database;
+import com.janbabak.noqlbackend.model.entity.User;
+import com.janbabak.noqlbackend.service.AuthenticationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import static com.janbabak.noqlbackend.error.exception.EntityNotFoundException.Entity.DATABASE;
@@ -25,41 +29,83 @@ import static com.janbabak.noqlbackend.error.exception.EntityNotFoundException.E
 public class DatabaseEntityService {
 
     private final DatabaseRepository databaseRepository;
+    private final UserRepository userRepository;
+    private final AuthenticationService authenticationService;
 
     /**
      * Find database by id.
      *
      * @param databaseId database identifier
      * @return database
-     * @throws EntityNotFoundException database of specified id not found.
+     * @throws EntityNotFoundException                                   database of specified id not found.
+     * @throws org.springframework.security.access.AccessDeniedException if user is not admin or owner of the database.
      */
     public Database findById(UUID databaseId) throws EntityNotFoundException {
         log.info("Get database by id={}.", databaseId);
 
-        return databaseRepository.findById(databaseId)
+        Database database = databaseRepository.findById(databaseId)
                 .orElseThrow(() -> new EntityNotFoundException(DATABASE, databaseId));
+
+        authenticationService.ifNotAdminOrSelfRequestThrowAccessDenied(database.getUserId());
+
+        return database;
     }
 
     /**
-     * Find all databases.
+     * Find all databases (from all users)
      *
      * @return list of databases
+     * @throws org.springframework.security.access.AccessDeniedException if user is not admin
      */
     public List<Database> findAll() {
+        return findAll(null);
+    }
+
+
+    /**
+     * Find all databases with filter.
+     *
+     * @param userId user identifier - filter by user id. When null - return all databases.
+     *               User can see only his databases. Admin can see all.
+     * @return list of databases
+     * @throws org.springframework.security.access.AccessDeniedException if user is not admin or owner of the database.
+     */
+    public List<Database> findAll(UUID userId) {
         log.info("Get all databases.");
 
-        return databaseRepository.findAll();
+        authenticationService.ifNotAdminOrSelfRequestThrowAccessDenied(userId);
+
+        return userId == null
+                ? databaseRepository.findAll()
+                : databaseRepository.findAllByUserId(userId);
     }
 
     /**
      * Create new database object - persist it.
      *
-     * @param database object to be saved
+     * @param request database data
      * @return saved object with id
-     * @throws DatabaseConnectionException if connection to the database failed.
+     * @throws DatabaseConnectionException                               if connection to the database failed.
+     * @throws org.springframework.security.access.AccessDeniedException if user is not admin or owner of the database.
      */
-    public Database create(Database database) throws DatabaseConnectionException {
+    public Database create(CreateDatabaseRequest request) throws DatabaseConnectionException, EntityNotFoundException {
         log.info("Create new database.");
+
+        authenticationService.ifNotAdminOrSelfRequestThrowAccessDenied(request.getUserId());
+
+        User user = userRepository.findById(request.getUserId()).orElseThrow(
+                () -> new EntityNotFoundException(EntityNotFoundException.Entity.USER, request.getUserId()));
+
+        Database database = Database.builder()
+                .name(request.getName())
+                .host(request.getHost())
+                .port(request.getPort())
+                .database(request.getDatabase())
+                .userName(request.getUserName())
+                .password(request.getPassword())
+                .engine(request.getEngine())
+                .user(user)
+                .build();
 
         DatabaseServiceFactory.getDatabaseDAO(database).testConnection();
 
@@ -70,10 +116,11 @@ public class DatabaseEntityService {
      * Update not null parameters of database.
      *
      * @param databaseId identifier of the database object to update
-     * @param data new data
+     * @param data       new data
      * @return updated object
-     * @throws EntityNotFoundException     database of specified id not found.
-     * @throws DatabaseConnectionException connection to the updated database failed.
+     * @throws EntityNotFoundException                                   database of specified id not found.
+     * @throws org.springframework.security.access.AccessDeniedException if user is not admin or owner of the database.
+     * @throws DatabaseConnectionException                               connection to the updated database failed.
      */
     public Database update(UUID databaseId, UpdateDatabaseRequest data)
             throws EntityNotFoundException, DatabaseConnectionException {
@@ -82,6 +129,8 @@ public class DatabaseEntityService {
 
         Database database = databaseRepository.findById(databaseId)
                 .orElseThrow(() -> new EntityNotFoundException(DATABASE, databaseId));
+
+        authenticationService.ifNotAdminOrSelfRequestThrowAccessDenied(database.getUserId());
 
         if (data.getName() != null) database.setName(data.getName());
         if (data.getHost() != null) database.setHost(data.getHost());
@@ -104,7 +153,12 @@ public class DatabaseEntityService {
     public void deleteById(UUID databaseId) {
         log.info("Delete database by id={}.", databaseId);
 
-        databaseRepository.deleteById(databaseId);
+        Optional<Database> database = databaseRepository.findById(databaseId);
+
+        if (database.isPresent()){
+            authenticationService.ifNotAdminOrSelfRequestThrowAccessDenied(database.get().getUserId());
+            databaseRepository.deleteById(databaseId);
+        }
     }
 
     /**
@@ -112,9 +166,10 @@ public class DatabaseEntityService {
      *
      * @param databaseId database identifier
      * @return database structure
-     * @throws EntityNotFoundException     database of specific id not found
-     * @throws DatabaseConnectionException connection to the database failed
-     * @throws DatabaseExecutionException  syntax error, ...
+     * @throws EntityNotFoundException                                   database of specific id not found
+     * @throws DatabaseConnectionException                               connection to the database failed
+     * @throws DatabaseExecutionException                                syntax error, ...
+     * @throws org.springframework.security.access.AccessDeniedException if user is not admin or owner of the database.
      */
     public DatabaseStructureDto getDatabaseStructureByDatabaseId(UUID databaseId)
             throws EntityNotFoundException, DatabaseConnectionException, DatabaseExecutionException {
@@ -122,22 +177,28 @@ public class DatabaseEntityService {
         Database database = databaseRepository.findById(databaseId)
                 .orElseThrow(() -> new EntityNotFoundException(DATABASE, databaseId));
 
+        authenticationService.ifNotAdminOrSelfRequestThrowAccessDenied(database.getUserId());
+
         return DatabaseServiceFactory.getDatabaseService(database).retrieveSchema().toDto();
     }
 
     /**
      * Get generated database create script by database id.
+     *
      * @param databaseId database identifier
      * @return create script
-     * @throws EntityNotFoundException database of specific id not found
-     * @throws DatabaseConnectionException cannot establish connection with the database
-     * @throws DatabaseExecutionException syntax error, ...
+     * @throws EntityNotFoundException                                   database of specific id not found
+     * @throws DatabaseConnectionException                               cannot establish connection with the database
+     * @throws DatabaseExecutionException                                syntax error, ...
+     * @throws org.springframework.security.access.AccessDeniedException if user is not admin or owner of the database.
      */
     public String getDatabaseCreateScriptByDatabaseId(UUID databaseId)
             throws EntityNotFoundException, DatabaseConnectionException, DatabaseExecutionException {
 
         Database database = databaseRepository.findById(databaseId)
                 .orElseThrow(() -> new EntityNotFoundException(DATABASE, databaseId));
+
+        authenticationService.ifNotAdminOrSelfRequestThrowAccessDenied(database.getUserId());
 
         return DatabaseServiceFactory.getDatabaseService(database).retrieveSchema().generateCreateScript();
     }
